requirements.txt
streamlit
google-generativeai
SpeechRecognition 
pyaudio
